Add str field to t_message struct and enhance character handling in server

+++ src/server.c
@@ -6,7 +6,7 @@
 /*   By: halmuhis <halmuhis@student.42amman.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/12 11:51:39 by halmuhis          #+#    #+#             */
-/*   Updated: 2025/02/14 20:16:18 by halmuhis         ###   ########.fr       */
+/*   Updated: 2025/02/16 20:59:01 by halmuhis         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,9 +17,23 @@ static t_message	g_msg;
 static void	print_char(void)
 {
 	if (g_msg.c == '\0')
-		printf("\nMessage received!\n");
+	{
+		ft_putstr_fd(g_msg.str, 1);
+		ft_putstr_fd("\n", 1);
+		free(g_msg.str);
+		g_msg.str = 0;
+	}
 	else
-		printf("%c", g_msg.c);
+	{
+		if (g_msg.str == 0)
+		{
+			g_msg.str = malloc(2);
+			g_msg.str[0] = g_msg.c;
+			g_msg.str[1] = 0;
+		}
+		else
+			ft_append(&g_msg.str, g_msg.c);
+	}
 	g_msg.c = 0;
 	g_msg.bits = 0;
 }
@@ -27,15 +41,18 @@ static void	print_char(void)
 static void	handle_signal(int sig, siginfo_t *info, void *context)
 {
 	(void)context;
-	if (sig == SIGUSR1)
-		g_msg.c = (g_msg.c << 1);
-	else if (sig == SIGUSR2)
-		g_msg.c = (g_msg.c << 1) | 1;
-	g_msg.bits++;
-	if (g_msg.bits == 8)
-		print_char();
-	if (kill(info->si_pid, SIGUSR1) == -1)
-		exit(EXIT_FAILURE);
+	if (sig == SIGUSR1 || sig == SIGUSR2)
+	{
+		if (sig == SIGUSR1)
+			g_msg.c = (g_msg.c << 1);
+		else if (sig == SIGUSR2)
+			g_msg.c = (g_msg.c << 1) | 1;
+		g_msg.bits++;
+		if (g_msg.bits == 8)
+			print_char();
+		if (kill(info->si_pid, SIGUSR1) == -1)
+			exit(EXIT_FAILURE);
+	}
 }
 
 static void	setup_signals(void)
@@ -58,8 +75,9 @@ int	main(void)
 	g_msg.c = 0;
 	g_msg.bits = 0;
 	setup_signals();
-	printf("Server PID: %d\n", getpid());
-	printf("Waiting for messages...\n");
+	ft_putstr_fd("Server PID: ", 1);
+	ft_putnbr_fd(getpid(), 1);
+	ft_putstr_fd("\n", 1);
 	while (1)
 		pause();
 	return (0);
